app/requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
colorama==0.4.6
coverage==7.6.1
Django==5.0.8
djangorestframework==3.15.2
model-bakery==1.18.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
sqlparse==0.5.1
tzdata==2024.1
